pyproject.toml
[build-system]
requires = ["setuptools>=61", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "dask-gateway-dashboard"
version = "0.0.1.dev"
dynamic = ["readme", "dependencies"]
description = "Serve a page that lists dask gateway clusters"
authors = [{ name = "Min RK", email = "benjaminrk@gmail.com" }]
keywords = ["Dask"]
license = { text = "BSD-3-Clause" }
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 1 - Planning",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: BSD License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
]

[project.urls]
Homepage = "https://github.com/minrk/dask-gateway-dashboard"
Source = "https://github.com/minrk/dask-gateway-dashboard"
Tracker = "https://github.com/minrk/dask-gateway-dashboard/issues"

[project.optional-dependencies]
test = [
  "dask-gateway-server[local]",
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "pytest-playwright-asyncio",
]


[tool.setuptools]
zip-safe = false
license-files = ["LICENSE"]
include-package-data = true
py-modules = ["dask_gateway_dashboard"]

# dynamic sources of metadata in other files
[tool.setuptools.dynamic]
readme = { file = "README.md", content-type = "text/markdown" }
dependencies = { file = "requirements.in" }

# setuptools_scm needs a section to be present
[tool.setuptools_scm]
# we don't actually use setuptools_scm for versions,
# only the file-finder
fallback_version = "0.0.0"

[tool.pytest.ini_options]
addopts = "--verbose --color=yes --durations=10 --browser firefox -p no:pytest-playwright"
testpaths = ["tests"]
asyncio_mode = "auto"
# needs to be 'session' for pytest-playwright
asyncio_default_fixture_loop_scope = "session"


[tool.ruff.format]

[tool.ruff.lint]
ignore = []
select = [
  "E9", # syntax
  "I",  # isort
  "UP", # pyupgrade
  "F",  # flake8
]

# tbump for tagging releases
[tool.tbump]
github_url = "https://github.com/minrk/dask-gateway-dashboard"

[tool.tbump.version]
current = "0.0.1.dev"

regex = '''
  (?P<major>\d+)
  \.
  (?P<minor>\d+)
  \.
  (?P<patch>\d+)
  (?P<pre>((a|b|rc)\d+)|)
  \.?
  (?P<dev>(?<=\.)dev\d*|)
  '''

[tool.tbump.git]
message_template = "Bump to {new_version}"
tag_template = "{new_version}"

# For each file to patch, add a [[tool.tbump.file]] config
# section containing the path of the file, relative to the
# pyproject.toml location.

[[tool.tbump.file]]
src = "pyproject.toml"
search = 'version = "{current_version}"'
